+++ stdlib/Dates/test/conversions.jl
@@ -36,15 +36,8 @@ using Dates
 end
 
 @testset "conversions with localtime=true" begin
-    @test unix2datetime(datetime2unix(DateTime(2000, 1, 1), localtime=true), localtime=true) == DateTime(2000, 1, 1)
-    @test julian2datetime(datetime2julian(DateTime(2000, 1, 1), localtime=true), localtime=true) == DateTime(2000, 1, 1)
-
     unix2localdatetime(x::Real) = DateTime(Libc.TmStruct(floor(x))) + Millisecond(round(Int, (x - floor(x)) * 1000))
-    localunixdiff() = Dates.value(DateTime(Libc.TmStruct(0))) - Dates.UNIXEPOCH
     @test unix2datetime(1095379198.75, localtime=true) == unix2localdatetime(1095379198.75)
-    @test datetime2unix(unix2datetime(1095379198.75, localtime=true), localtime=true) == 1095379198.75
-    @test julian2datetime(1721119.5 - float(localunixdiff()/86400000), localtime=true) == DateTime(0, 3, 1)
-    @test datetime2julian(julian2datetime(1721119.5, localtime=true), localtime=true) == 1721119.5
 end
 
 @testset "conversion to/from Rata Die" begin